pyproject.toml
[project]
name = "deepchat"
version = "0.1.0"
description = ""
authors = [
    {name = "Bektur",email = "bekturkochorbaev64@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "annotated-types (>=0.7.0,<0.8.0)",
    "anyio (>=4.8.0,<5.0.0)",
    "bcrypt (>=4.2.1,<5.0.0)",
    "build (>=1.2.2.post1,<2.0.0)",
    "cachecontrol (>=0.14.2,<0.15.0)",
    "certifi (>=2025.1.31,<2026.0.0)",
    "charset-normalizer (>=3.4.1,<4.0.0)",
    "click (>=8.1.8,<9.0.0)",
    "colorama (>=0.4.6,<0.5.0)",
    "distlib (>=0.3.9,<0.4.0)",
    "dnspython (>=2.7.0,<3.0.0)",
    "dulwich (>=0.22.7,<0.23.0)",
    "ecdsa (>=0.19.0,<0.20.0)",
    "email-validator (>=2.2.0,<3.0.0)",
    "fastapi (>=0.115.8,<0.116.0)",
    "fastapi-cli (>=0.0.7,<0.0.8)",
    "fastjsonschema (>=2.21.1,<3.0.0)",
    "filelock (>=3.17.0,<4.0.0)",
    "findpython (>=0.6.2,<0.7.0)",
    "h11 (>=0.14.0,<0.15.0)",
    "httpcore (>=1.0.7,<2.0.0)",
    "httptools (>=0.6.4,<0.7.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "idna (>=3.10,<4.0)",
    "installer (>=0.7.0,<0.8.0)",
    "jaraco-classes (>=3.4.0,<4.0.0)",
    "jaraco-context (>=6.0.1,<7.0.0)",
    "jaraco-functools (>=4.1.0,<5.0.0)",
    "jinja2 (>=3.1.5,<4.0.0)",
    "keyring (>=25.6.0,<26.0.0)",
    "markdown-it-py (>=3.0.0,<4.0.0)",
    "markupsafe (>=3.0.2,<4.0.0)",
    "mdurl (>=0.1.2,<0.2.0)",
    "more-itertools (>=10.6.0,<11.0.0)",
    "motor (>=3.7.0,<4.0.0)",
    "msgpack (>=1.1.0,<2.0.0)",
    "packaging (>=24.2,<25.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "pbs-installer (>=2025.2.12,<2026.0.0)",
    "pkginfo (>=1.12.1.2,<2.0.0.0)",
    "platformdirs (>=4.3.6,<5.0.0)",
    "pyasn1 (>=0.6.1,<0.7.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "pydantic-core (>=2.29.0,<3.0.0)",
    "pygments (>=2.19.1,<3.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "pymongo (>=4.11.1,<5.0.0)",
    "pyproject-hooks (>=1.2.0,<2.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "python-jose (>=3.3.0,<4.0.0)",
    "pywin32-ctypes (>=0.2.3,<0.3.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "rapidfuzz (>=3.12.1,<4.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "requests-toolbelt (>=1.0.0,<2.0.0)",
    "rich (>=13.9.4,<14.0.0)",
    "rich-toolkit (>=0.13.2,<0.14.0)",
    "shellingham (>=1.5.4,<2.0.0)",
    "six (>=1.17.0,<2.0.0)",
    "sniffio (>=1.3.1,<2.0.0)",
    "tomlkit (>=0.13.2,<0.14.0)",
    "trove-classifiers (>=2025.2.18.16,<2026.0.0.0)",
    "typer (>=0.15.1,<0.16.0)",
    "typing-extensions (>=4.12.2,<5.0.0)",
    "urllib3 (>=2.3.0,<3.0.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "virtualenv (>=20.29.2,<21.0.0)",
    "watchfiles (>=1.0.4,<2.0.0)",
    "websockets (>=15.0,<16.0)",
    "zstandard (>=0.23.0,<0.24.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
[tool.poetry]
package-mode = false
